requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic[email]